feat: moved the refresh token to a global component
A  src/app/GlobalApp.js
A  src/app/_app.js
M  src/app/doctoralCenter/admin/(home)/page.jsx
M  src/app/layout.jsx
